AC_Overlay-1.8
Added a lot of preliminary block updates in the wake of Lorien and for
greater advancement of the Beorning Areas. More reworks to come in
1.12, only ones added for now are ones needed for current projects.

Changelog:
-Glacier Ice updated
-Granite & its Variants updated
-Gneiss updated
-Rough Marble updated
-Niphredil added
-Hartstongue modified
-Greywacke & all Sandstone variants updated
-Big limestone slab updated
-Calcite updated
-Wood blocks all/mostly updated due to Beorning projects
-Red sandstone bricks updated
-Seashell elf wall updated
-Sandstone Bigslab updated
-drystone cobble wall updated
-Beech log updated
-Grey Ivory Pillar updated for Lorien
-Grey painted planks updated for Lorien
-White painted horizontal planks updated for Lorien
-Paintings updated to most recent versions from the discord, no more
ugly numbers in Rivendell and some Lorien paintings are now fully in
the game.

*Parts of this are from Monster’s texture rework, and parts are of my
own, and the painting team’s, addition. Hopefully projects can go more
smoothly now.